pyproject.toml
[project]
name = "thermal-printer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-escpos>=3.1",
]

[project.scripts]
ticket = "main:main"

[tool.uv]
package = true
